增加知识文档中心，增加分析图标接口，优化权限控制等等

+++ yudao-module-system/yudao-module-system-biz/src/main/java/cn/iocoder/yudao/module/system/service/impl/WenXunSpiderConfigServiceImpl.java
@@ -65,7 +65,7 @@ public class WenXunSpiderConfigServiceImpl implements WenXunSpiderConfigService 
 
                         .status(1L)
                         .itemXpath(createReqVO.getItemXpath())
-                        .deptId(createReqVO.getDeptId())
+                        .deptId((long)createReqVO.getDeptId())
                         .creator(SecurityFrameworkUtils.getLoginUserNickname())
                         .updater(SecurityFrameworkUtils.getLoginUserNickname())
                         .build();
